fix build/warning issues surfaced by new Ubuntu CI toolchain (#5736)
* fix problem 208 C++ solution by ignoring `-Wwrite-strings` warning


* fix problem 806 C solution by adding missing `<string>` header


* fix problem 821 C++ solution by adding missing `<string>` header


* fix problem 893 C++ solution by adding missing `<algorithm>` header


* fix problem 999 solutions by adding `fallthrough` attribute


* fix problem 1356 C solution by adding missing `<cstring>` header


* fix problem 1385 C++ solution by adding missing `<algorithm>` header


* fix problem 1461 C solution by adding missing `<string>` and `<stdint.h>` headers


* fix problem 1665 C++ solution by adding missing `<algorithm>` header


* fix problem 1769 C++ solution by adding missing `<string>` header


* fix problem 1833 C++ solution by adding missing `<algorithm>` header


* fix problem 1848 C++ solution by adding missing `<cstdlib>` header


* fix problem 1970 C++ solution by adding missing `leftParent` and `rightParent` field initializers


* fix problem 2452 C solution by adding missing `<string>` header


* fix problem 2506 C++ solution by adding missing `<algorithm>` header


* fix problem 2515 C solution by adding missing `<string>` header


* fix problem 2586 C solution by adding missing `<string>` header


* fix problem 3531 C++ solution by adding missing `<limits>` header


* fix problem 3643 C solution by adding missing `<cstring>` header


* fix problem 3731 C++ solution by adding missing `<algorithm>` header


* fix problem 3803 C++ solution by adding missing `<cstdint>` header and casting `i` to `int` to avoid sign-conversion warning


* fix problem 874 C++ solution by adding missing `<unordered_map>` header


* fix problem 811 test cases by reordering expected output


* fix problem 2135 C solution by moving `startMasksExist` array allocation to heap


---------

+++ old-problems/3803/solution.cpp
@@ -1,4 +1,5 @@
 #include <bit>
+#include <cstdint>
 #include <string>
 
 class Solution {
@@ -8,7 +9,7 @@ class Solution {
     std::uint32_t distincts{0};
     for (std::size_t i{0}; i < s.size(); ++i) {
       distincts |= 1 << (s[i] - 'a');
-      if (std::popcount(distincts) == (i + 1) % 3) ++count;
+      if (std::popcount(distincts) == (static_cast<int>(i) + 1) % 3) ++count;
     }
     return count;
   }
